Add files via upload
It is a simple and fun project using ‘’python’’. It will ask the user name of a particular websites URL. Then it will scripe that website and collect all the link there is in that web site. I used ‘’tkinter” for GUI design and ‘’bs4’’ for simplifying the HTML code.